test(e2e): organize the product e2e suite per surface
Root e2e/ is product-level tests through a user-facing surface, one
subdirectory per surface: e2e/web/ holds today's browser suites, a future
surface (bot, CLI) adds its own e2e/<surface>/ instead of competing for
the root. Shared boot infra (db/model servers, fixtures, auth helpers)
moves to e2e/support/. App-scoped suites live with their app
(apps/api/e2e) — they never belonged here.

Verified: playwright test --list collects all 19 tests in 6 files.
