+++ e2e/web/chat/model-context-transparency.spec.ts
@@ -8,7 +8,7 @@
  * to rebuild after every hidden context artifact exists.
  */
 
-import { expect, test } from "../fixtures";
+import { expect, test } from "../../support/fixtures";
 import { seedCompaction } from "./seed-compaction";
 
 const ANSWER = "Mocked answer from the e2e model server.";
